Move shell to module
